Task4

+++ Tyuiu.AlshinAF.Sprint1.Task4.V23/Program.cs
@@ -22,11 +22,11 @@ namespace Tyuiu.AlshinAF.Sprint1.Task4.V23
             Console.WriteLine("* ИСХОДНЫЕ ДАННЫЕ:                                                        *");
             Console.WriteLine("***************************************************************************");
 
-            int x, y;
+            double x, y;
             Console.WriteLine("Введите значение X:");
-            x = Convert.ToInt32(Console.ReadLine());
+            x = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("Введите значение Y:");
-            y = Convert.ToInt32(Console.ReadLine());
+            y = Convert.ToDouble(Console.ReadLine());
             Console.WriteLine("***************************************************************************");
             Console.WriteLine("* РЕЗУЛЬТАТ:                                                              *");
             Console.WriteLine("***************************************************************************");
